fix: broken test
