+++ packages/marketplace/src/components/pages/__tests__/developer-home.test.tsx
@@ -5,7 +5,7 @@ import { mount } from 'enzyme'
 import configureStore from 'redux-mock-store'
 import appState from '@/reducers/__stubs__/app-state'
 import Routes from '@/constants/routes'
-import DeveloperHome, { handleOnCardClick, handleOnChange } from '../developer-home'
+import DeveloperHome, { handleOnCardClick, handleOnChange, onShowSubmitAppModal } from '../developer-home'
 import { AppSummaryModel } from '@/types/marketplace-api-schema'
 import { getMockRouterProps } from '@/utils/mock-helper'
 import routes from '@/constants/routes'
@@ -18,6 +18,7 @@ describe('Login', () => {
     const mockStore = configureStore()
     store = mockStore(appState)
   })
+
   it('should match a snapshot', () => {
     expect(
       mount(
@@ -29,6 +30,7 @@ describe('Login', () => {
       ),
     ).toMatchSnapshot()
   })
+
   describe('handleOnCardClick', () => {
     it('should run correctly', () => {
       const mockAppSummary: AppSummaryModel = {
@@ -45,6 +47,22 @@ describe('Login', () => {
       const fn = handleOnChange(history)
       fn(1)
       expect(history.push).toBeCalledWith(`${routes.DEVELOPER_MY_APPS}?page=${1}`)
+    })
+  })
+
+  describe('onShowSubmitAppModal', () => {
+    it('should set as visible', () => {
+      const setSubmitAppModalVisible = jest.fn()
+      onShowSubmitAppModal(setSubmitAppModalVisible)()
+      expect(setSubmitAppModalVisible).toHaveBeenCalledWith(true)
+    })
+  })
+
+  describe('onCloseSubmitAppModal', () => {
+    it('should set visible to false', () => {
+      const setSubmitAppModalVisible = jest.fn()
+      onShowSubmitAppModal(setSubmitAppModalVisible)()
+      expect(setSubmitAppModalVisible).toHaveBeenCalledWith(false)
     })
   })
 })
