+++ packages/marketplace/src/components/pages/developer-home.tsx
@@ -10,6 +10,7 @@ import { SandboxPopUp } from '@/components/ui/sandbox-pop-up'
 import { AppSummaryModel } from '@reapit/foundations-ts-definitions'
 import { getParamValueFromPath } from '@/utils/client-url-params'
 import Routes from '@/constants/routes'
+import { SubmitAppWizardModal } from '../ui/submit-app-wizard'
 
 export const handleOnCardClick = (history: History) => (app: AppSummaryModel) => {
   history.push(`${Routes.DEVELOPER_MY_APPS}/${app.id}`)
@@ -18,15 +19,16 @@ export const handleOnCardClick = (history: History) => (app: AppSummaryModel) =>
 export const handleOnChange = (history: History) => (page: number) =>
   history.push(`${Routes.DEVELOPER_MY_APPS}?page=${page}`)
 
-export const handleNewAppPage = (history: History) => () => {
-  history.push(Routes.SUBMIT_APP)
-}
+export const onShowSubmitAppModal = (setSubmitAppModalVisible: React.Dispatch<React.SetStateAction<boolean>>) => () =>
+  setSubmitAppModalVisible(true)
 
-export type DeveloperProps = {}
+export const onCloseSubmitAppModal = (setSubmitAppModalVisible: React.Dispatch<React.SetStateAction<boolean>>) => () =>
+  setSubmitAppModalVisible(false)
 
-export const DeveloperHome: React.FunctionComponent<DeveloperProps> = () => {
+export const DeveloperHome: React.FC = () => {
   const history = useHistory()
   const developerState = useSelector(selectDeveloper)
+  const [submitAppModalVisible, setSubmitAppModalVisible] = React.useState<boolean>(false)
 
   let pageNumber = 1
 
@@ -48,27 +50,24 @@ export const DeveloperHome: React.FunctionComponent<DeveloperProps> = () => {
 
   return (
     <ErrorBoundary>
-      <div id="page-developer-home-container">
-        <Section className="justify-between items-center" isFlex>
-          <H3 className="mb-0">My Apps</H3>
-          <Button onClick={handleNewAppPage(history)} type="button" variant="primary">
-            Create new app
-          </Button>
-        </Section>
-        <AppList
-          list={list}
-          loading={loading}
-          onCardClick={handleOnCardClick(history)}
-          infoType="DEVELOPER_APPS_EMPTY"
+      <Section className="justify-between items-center" isFlex>
+        <H3 className="mb-0">My Apps</H3>
+        <Button onClick={onShowSubmitAppModal(setSubmitAppModalVisible)} type="button" variant="primary">
+          Create new app
+        </Button>
+        <SubmitAppWizardModal
+          visible={submitAppModalVisible}
+          afterClose={onCloseSubmitAppModal(setSubmitAppModalVisible)}
         />
-        <Pagination
-          totalCount={totalCount}
-          pageSize={pageSize}
-          pageNumber={pageNumber}
-          onChange={handleOnChange(history)}
-        ></Pagination>
-        <SandboxPopUp loading={loading} />
-      </div>
+      </Section>
+      <AppList list={list} loading={loading} onCardClick={handleOnCardClick(history)} infoType="DEVELOPER_APPS_EMPTY" />
+      <Pagination
+        totalCount={totalCount}
+        pageSize={pageSize}
+        pageNumber={pageNumber}
+        onChange={handleOnChange(history)}
+      ></Pagination>
+      <SandboxPopUp loading={loading} />
     </ErrorBoundary>
   )
 }
