+++ packages/marketplace/src/components/ui/__tests__/app-list.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { shallow } from 'enzyme'
 
-import { AppList, AppListProps, onCloseSubmitAppModal, onShowSubmitAppModal } from '../app-list'
+import { AppList, AppListProps } from '../app-list'
 import { appsDataStub } from '@/sagas/__stubs__/apps'
 import { Loader, GridFourColItem, GridThreeColItem, Helper } from '@reapit/elements'
 import { AppSummaryModel } from '@reapit/foundations-ts-definitions'
@@ -12,31 +12,16 @@ const props: AppListProps = {
   loading: false,
   onCardClick: jest.fn(),
   onSettingsClick: jest.fn(),
-  title: 'Title',
   infoType: 'CLIENT_APPS_EMPTY',
 }
 
 describe('AppList', () => {
-  test('onShowSubmitAppModal should run correctly', () => {
-    const setSubmitAppModalVisible = jest.fn()
-    onShowSubmitAppModal(setSubmitAppModalVisible)()
-    expect(setSubmitAppModalVisible).toHaveBeenCalledWith(true)
-  })
-  test('onCloseSubmitAppModal should run correctly', () => {
-    const setSubmitAppModalVisible = jest.fn()
-    onCloseSubmitAppModal(setSubmitAppModalVisible)()
-    expect(setSubmitAppModalVisible).toHaveBeenCalledWith(false)
-  })
   it('should match a snapshot', () => {
     expect(shallow(<AppList {...props} />)).toMatchSnapshot()
   })
 
   it('should match a snapshot when use empty infoType', () => {
     expect(shallow(<AppList {...props} infoType={''} />)).toMatchSnapshot()
-  })
-
-  it('should match a snapshot when use has header button', () => {
-    expect(shallow(<AppList {...props} hasSubmitButton />)).toMatchSnapshot()
   })
 
   it('should show loading', () => {
