+++ packages/marketplace/src/components/ui/app-list.tsx
@@ -2,8 +2,7 @@ import * as React from 'react'
 import { AppSummaryModel } from '@reapit/foundations-ts-definitions'
 import AppCard from './app-card'
 import styles from '@/styles/blocks/app-list.scss?mod'
-import { Loader, H3, InfoType, GridFourCol, GridThreeColItem, Helper, infoText, Button } from '@reapit/elements'
-import { SubmitAppWizardModal } from '../ui/submit-app-wizard'
+import { Loader, InfoType, GridFourCol, GridThreeColItem, Helper, infoText } from '@reapit/elements'
 
 export type AppListProps = {
   list: AppSummaryModel[]
@@ -11,39 +10,6 @@ export type AppListProps = {
   onCardClick?: (app: AppSummaryModel) => void
   onSettingsClick?: (app: AppSummaryModel) => void
   infoType: InfoType
-  hasSubmitButton?: boolean
-}
-
-export const onShowSubmitAppModal = (setSubmitAppModalVisible: React.Dispatch<React.SetStateAction<boolean>>) => () => {
-  setSubmitAppModalVisible(true)
-}
-
-export const onCloseSubmitAppModal = (
-  setSubmitAppModalVisible: React.Dispatch<React.SetStateAction<boolean>>,
-) => () => {
-  setSubmitAppModalVisible(false)
-}
-
-export const renderHeader = ({ hasSubmitButton, title }: { hasSubmitButton: boolean; title: string | undefined }) => {
-  const [submitAppModalVisible, setSubmitAppModalVisible] = React.useState<boolean>(false)
-
-  if (!title) return null
-  if (!hasSubmitButton) {
-    return <H3>{title}</H3>
-  } else {
-    return (
-      <div className={styles.headerHasButton}>
-        <H3>{title}</H3>
-        <Button onClick={onShowSubmitAppModal(setSubmitAppModalVisible)} type="button" variant="primary">
-          Create new app
-        </Button>
-        <SubmitAppWizardModal
-          visible={submitAppModalVisible}
-          afterClose={onCloseSubmitAppModal(setSubmitAppModalVisible)}
-        />
-      </div>
-    )
-  }
 }
 
 export const AppList: React.FunctionComponent<AppListProps> = ({
